requirements.txt
discord.py-self
aiohttp
pyqt5